# processing/openapi.yml
openapi: 3.0.3
info:
  title: To-Do Checklist Processing API
  version: "1.0.0"
  description: Computes statistics from events stored in the storage service.

paths:
  /health: # A1: health check endpoint
    get:
      summary: Health check
      operationId: app.health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /stats:
    get:
      summary: gets the event stats
      operationId: app.get_event_stats
      responses:
        '200':
          description: sucessfully returned the storage stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
  /stats:
    get:
      summary: Get checklist stats
      operationId: app.get_stats
      description: Returns computed event statistics from the local stats file.
      responses:
        '200':
          description: Stats returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistStats'
        '404':
          description: Stats do not exist yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'

components:
  schemas:
    HealthResponse: # A1: health check response schema
      type: object
      properties:
        status:
          type: string
          example: Processing is healthy
    Stats:
      type: object
      required:
        - num_temp
        - num_precip
      properties:
        num_temp:
          type: integer
          example: 100
        num_precip:
          type: integer
          example: 100
    ChecklistStats:
      type: object
      required:
        - num_checklist_item_events
        - num_checklist_summary_events
        - max_estimated_mins
        - max_items_completed
        - last_updated
      properties:
        num_checklist_item_events:
          type: integer
         
        num_checklist_summary_events:
          type: integer
          
        max_estimated_mins:
          type: integer
          nullable: true
         
        max_items_completed:
          type: integer
          nullable: true
         
        last_updated:
          type: string
          format: date-time
          nullable: true
      

    ErrorMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          